fixed gql
