+++ src/graphql/GetRepositoryQuery.graphql
@@ -1,0 +1,9 @@
+query GetRepositoryQuery($owner: String!, $repo: String!) {
+  repository(owner: $owner, name: $repo) {
+    id
+    defaultBranchRef {
+      prefix
+      name
+    }
+  }
+}
